requirements.txt
boto3==1.21.29
fastapi==0.75.0
imgaug==0.4.0
numpy==1.20.0
paddlenlp==2.5.1
paddlepaddle==2.4.2
paddleocr==2.6
polygon3==3.0.9.1
opencv-python==4.6.0.66
pandas==1.3.0
pillow==9.4.0
motor==2.5.1
pyyaml==6.0
uvicorn==0.23.2
python-multipart==0.0.5